Make cache_dir a required param when initializing LiveFileSourceSSH
`cache_dir=None` seems to imply that `cache_dir` can be omitted, or that the default value of `None` will mean no cache dir is used. 

But actually the class fails to initialize because line 185 expects `cache_dir` to be non-null. 

This change requires `cache_dir` to be specified so that the class does not fail to initialize during runtime.